requirements.txt
asgiref==3.3.2
astroid==2.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
crontab==0.22.9
Django==3.2.1
django-braces==1.14.0
django-cors-headers==3.4.0
django-discover-runner==1.0
django-filter==2.3.0
django-hitcount==1.3.2
django-modeltranslation==0.15.1
django-resized
django-role-permissions==3.1.0
djangorestframework==3.11.1
drf-nested-routers==0.93.3
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
isort==5.8.0
itypes==1.2.0
Jinja2==3.0.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
packaging==20.9
Pillow==9.2.0
psycopg2-binary==2.8.6
pylint==2.6.0
pyparsing==2.4.7
pytz==2020.1
requests==2.25.1
sentry-sdk==1.43.0
six==1.15.0
sorl-thumbnail==12.9.0
sqlparse==0.3.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.18
vcversioner==2.16.0.0
wrapt==1.12.1
zope.interface==5.1.0
django-ckeditor==6.0.0
django-image-optimizer==1.0.1
django-rosetta==0.9.7
django-recaptcha==3.0.0
django-multiselectfield==0.1.12
xlwt==1.3.0
django-js-asset==1.2.2
typing_extensions==4.7.1
factory-boy==3.3.0
python-environ==0.4.54
django-jazzmin==2.6.0
phonenumbers==8.13.25
django-phonenumber-field==7.2.0
djangorestframework-simplejwt==5.3.0
django-redis==5.4.0
gunicorn
django-debug-toolbar==3.3.0
celery>=5.0.0,<5.4.0
